refactor: move APIEnums into api module
APIEnums are concerned with API code and should not be used by
the engine core. This refactoring removes the last dependencies
of engine code on APIEnums and moves the APIEnums class into
the api module.
