+++ ors-api/src/main/java/org/heigit/ors/api/controllers/SnappingAPI.java
@@ -35,7 +35,7 @@ import org.heigit.ors.api.responses.snapping.json.JsonSnappingResponse;
 import org.heigit.ors.api.services.SnappingService;
 import org.heigit.ors.api.util.AppConfigMigration;
 import org.heigit.ors.exceptions.*;
-import org.heigit.ors.routing.APIEnums;
+import org.heigit.ors.api.APIEnums;
 import org.heigit.ors.snapping.SnappingErrorCodes;
 import org.heigit.ors.snapping.SnappingResult;
 import org.springframework.core.convert.ConversionFailedException;
